rvgpu-renderer: Fix artefacts under GNOME
When framebuffer has ALPHA channel and the window created
under GNOME wayland, there are artifacts.
Probably this should be fixed if we combine it with
the support for "-a enable translucent mode on Wayland" flag.

Also this patch implements choosing EGL config taking into
account the native format required by GBM layer.